Fix an issue with the stream crashing in `Sound`
This was due to the user-friendly delegate/callback getting GC'd by the
C# runtime, right when portaudio needed to call it.

This fix (unfortanely) gets rid of that typesafe callback.  To
compromise (and make the programmer's life easier), there is a new
method to convert a `userDataPtr`'s `IntPtr` (in the port audio
callbacks), to the actual type of the User data.  Take a look at the
`Sinewave` example to see what I'm talking about.

This also includes some documentation fixups