Switch to using Vcpkg to get third party deps for Bassoon
By this, I mean the native libraries for libsndfile and portaudio.  That
CMake file was nasty and started to break on Windows.  Using Vcpkg is
way more elegant and nice.

The setup.py file is also more logical IMO, even if things are a tad
more verbose than standard bash scripting.  I like it.